feat(about): copy your environment (#649)
